requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.1
distro==1.9.0
exceptiongroup==1.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
oauthlib==3.2.2
openai==1.76.0
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
requests==2.32.3
requests-oauthlib==2.0.0
schedule==1.2.2
sniffio==1.3.1
tqdm==4.67.1
tweepy==4.15.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
